move zap to CHIP build system
