+++ src/app/chip-zcl/Makefile.am
@@ -1,0 +1,34 @@
+#
+#    Copyright (c) 2020 Project CHIP Authors
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+#
+
+#
+#    Description:
+#      This file exports the CHIP ZCL Application layer API headers
+#       for install and dist targets
+#
+
+include $(abs_top_nlbuild_autotools_dir)/automake/pre.am
+
+chip_zcldir = $(includedir)/chip-zcl
+
+chip_zcl_HEADERS                              = \
+    chip-zcl-buffer.h                           \
+    chip-zcl-codec.h                           	\
+    chip-zcl-struct.h                          	\
+    chip-zcl.h                                 	\
+    $(NULL)
+
+include $(abs_top_nlbuild_autotools_dir)/automake/post.am
