+++ src/app/plugin/test-unit/ChipZclUnitTests.c
@@ -1,0 +1,34 @@
+/**
+ *
+ *    Copyright (c) 2020 Project CHIP Authors
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+#include <stdio.h>
+#include <stdlib.h>
+
+#include "ChipZclUnitTests.h"
+
+int main(int argc, char ** argv)
+{
+    printf("%s: === start ===\n", argv[0]);
+    testClusterServerBasic();
+    testClusterServerIdentify();
+    testClusterServerLevelControl();
+    testClusterServerOnOff();
+    testCodecSimple();
+    testCoreDataModel();
+    testCoreMessageDispatch();
+    printf("%s: === end ===\n", argv[0]);
+    return 0;
+}
