+++ src/app/plugin/test-unit/ChipZclUnitTests.h
@@ -1,0 +1,28 @@
+/**
+ *
+ *    Copyright (c) 2020 Project CHIP Authors
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+#ifndef CHIP_ZCL_TEST_UNIT
+#define CHIP_ZCL_TEST_UNIT
+
+int testClusterServerBasic();
+int testClusterServerIdentify();
+int testClusterServerLevelControl();
+int testClusterServerOnOff();
+int testCodecSimple();
+int testCoreDataModel();
+int testCoreMessageDispatch();
+
+#endif // CHIP_ZCL_TEST_UNIT
