+++ src/app/plugin/test-unit/core-message-dispatch-test.c
@@ -22,7 +22,11 @@
  *
  */
 
-#include "test-unit.h"
+#include "ChipZclUnitTests.h"
+
+#include <chip-zcl.h>
+
+#include "gen.h"
 
 // Main command parsing controller.
 ChipZclStatus_t chipZclCommandParse(ChipZclCommandContext_t * context);
